Object created from Program.cs to show in console screen

+++ MyFirstOOPCode-Saturday/MyFirstOOPCode-Saturday/Employee.cs
@@ -16,7 +16,7 @@ namespace MyFirstOOPCode_Saturday
         #endregion
 
         #region Methods
-        public abstract int GetValueToPay();
+        public abstract decimal GetValueToPay();
 
         public override string ToString()
         {
